+++ api/V0ComposeKillTask.go
@@ -7,11 +7,11 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// V0ComposeKillTask will kill a task from a service from a specific project
+// V0ComposeRestartTask will restart a task
 // example:
-// curl -X DELETE http://user:password@127.0.0.1:10000/api/compose/v0/{projectname}/{servicename}/{taskid}
+// curl -X DELETE http://user:password@127.0.0.1:10000/api/compose/v0/tasks/{taskid}
 func (e *API) V0ComposeKillTask(w http.ResponseWriter, r *http.Request) {
-	logrus.WithField("func", "api.V0ComposeKillTask").Debug("Kill Task")
+	logrus.WithField("func", "api.V0ComposeKillTask").Debug("Kill Mesos Task")
 
 	vars := mux.Vars(r)
 	auth := e.CheckAuth(r, w)
@@ -24,24 +24,24 @@ func (e *API) V0ComposeKillTask(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if vars["project"] == "" || vars["servicename"] == "" || vars["taskid"] == "" {
+	if vars["taskid"] == "" {
 		w.WriteHeader(http.StatusNotAcceptable)
 		return
 	}
 
-	project := vars["project"]
-	servicename := vars["servicename"]
 	taskID := vars["taskid"]
 
-	key := e.Redis.GetRedisKey(e.Config.PrefixTaskName + ":" + project + ":" + servicename + ":" + taskID)
+	task := e.Redis.GetTaskFromTaskID(taskID)
 
-	task := e.Mesos.DecodeTask(key)
-	if task.TaskID == taskID {
-		task.State = "__KILL"
-		task.Restart = "no"
-		e.Redis.SaveTaskRedis(task)
-		e.Mesos.SuppressFramework()
-
-		logrus.WithField("func", "api.V0ComposeKillTask").Info("Kill Task: " + e.Config.PrefixTaskName + ":" + project + ":" + servicename + ":" + taskID)
+	if task.TaskID == "" {
+		logrus.WithField("func", "api.V0ComposeKillTask").Errorf("Could not find TaskID (%s)", taskID)
+		w.WriteHeader(http.StatusNotAcceptable)
+		return
 	}
+
+	logrus.WithField("func", "api.V0ComposeKillTask").Infof("Kill Task (%s)", taskID)
+
+	task.State = "__KILL"
+	task.Restart = "no"
+	e.Redis.SaveTaskRedis(task)
 }