+++ api/api.go
@@ -52,13 +52,14 @@ func (e *API) Commands() *mux.Router {
 	rtr := mux.NewRouter()
 	rtr.HandleFunc("/api/compose/versions", e.Versions).Methods("GET")
 	rtr.HandleFunc("/api/compose/v0/tasks", e.V0ShowAllTasks).Methods("GET")
+	rtr.HandleFunc("/api/compose/v0/tasks/{taskid}", e.V0ComposeKillTask).Methods("DELETE")
+	rtr.HandleFunc("/api/compose/v0/tasks/{taskid}/restart", e.V0ComposeRestartTask).Methods("PUT")
 	rtr.HandleFunc("/api/compose/v0/framework/reregister", e.V0FrameworkReRegister).Methods("PUT")
 	rtr.HandleFunc("/api/compose/v0/framework/suppress", e.V0FrameworkSuppress).Methods("PUT")
 	rtr.HandleFunc("/api/compose/v0/{project}", e.V0ComposePush).Methods("PUT")
 	rtr.HandleFunc("/api/compose/v0/{project}", e.V0ComposeUpdate).Methods("UPDATE")
 	rtr.HandleFunc("/api/compose/v0/{project}/{servicename}", e.V0ComposeKillService).Methods("DELETE")
 	rtr.HandleFunc("/api/compose/v0/{project}/{servicename}/restart", e.V0ComposeRestartService).Methods("PUT")
-	rtr.HandleFunc("/api/compose/v0/{project}/{servicename}/{taskid}", e.V0ComposeKillTask).Methods("DELETE")
 
 	return rtr
 }