+++ redis/redis.go
@@ -100,6 +100,27 @@ func (e *Redis) GetTaskFromEvent(update *mesosproto.Event_Update) *cfg.Command {
 	return &cfg.Command{}
 }
 
+// GeTaskFromTaskID get out the task by a taskID
+func (e *Redis) GetTaskFromTaskID(taskID string) *cfg.Command {
+	// search matched taskid in redis and update the status
+	keys := e.GetAllRedisKeys(e.Prefix + ":*")
+	for keys.Next(e.CTX) {
+		// ignore redis keys if they are not mesos tasks
+		if e.CheckIfNotTask(keys) {
+			continue
+		}
+		// get the values of the current key
+		key := e.GetRedisKey(keys.Val())
+		task := e.Mesos.DecodeTask(key)
+
+		if task.TaskID == taskID {
+			return task
+		}
+	}
+
+	return &cfg.Command{}
+}
+
 // CountRedisKey will get back the count of the redis key
 func (e *Redis) CountRedisKey(pattern string, ignoreState string) int {
 	keys := e.GetAllRedisKeys(pattern)