+++ shell.nix
@@ -1,14 +1,23 @@
-{ pkgs ? import <nixpkgs> { } }:
+with import <nixpkgs> {};
 
-with pkgs;
+stdenv.mkDerivation {
+name = "go-env";
 
-mkShell {
-  buildInputs = [
+buildInputs = [
     go
     syft
     grype
-    docker-credential-helpers    
     docker
     trivy
-  ];
+];
+
+SOURCE_DATE_EPOCH = 315532800;
+PROJDIR = "${toString ./.}";
+S_NETWORK="host";
+
+shellHook = ''
+    export LD_LIBRARY_PATH="${pkgs.stdenv.cc.cc.lib}/lib"
+    export PATH=/tmp/bin:$PATH
+    mkdir /tmp/bin
+    '';
 }